Start logic for tutorial
This will be the main tutorial for the game; a simple step by step overlay to
teadch the user. The Rules page will supplement this with further explanation.

This will only show on a users first time using the app. Subsequent time they
will need to access it through the rules page.

- ph-tutorial directive.
- Start tutorial overlay design. This will animate when pages change, using the
  same logic the app does.
- Update go function to only open links if they aren't currently the active page.
- Disable logo clicks in tutorial mode.